+++ banco/dados_tb_indicador_classificacao.sql
@@ -1,0 +1,34 @@
+--
+-- PostgreSQL database dump
+--
+
+-- Dumped from database version 9.6.2
+-- Dumped by pg_dump version 9.6.2
+
+SET statement_timeout = 0;
+SET lock_timeout = 0;
+SET idle_in_transaction_session_timeout = 0;
+SET client_encoding = 'UTF8';
+SET standard_conforming_strings = on;
+SET check_function_bodies = false;
+SET client_min_messages = warning;
+SET row_security = off;
+
+SET search_path = dbmgi, pg_catalog;
+
+--
+-- Data for Name: tb_indicador_classificacao; Type: TABLE DATA; Schema: dbesusgestor; Owner: postgres
+--
+
+COPY tb_indicador_classificacao (co_indicador_classificacao, ds_indicador_classificacao) FROM stdin;
+1	Insumo
+2	Processo
+3	Resultado
+4	Produto
+5	Impacto
+\.
+
+
+--
+-- PostgreSQL database dump complete
+--