+++ client/src/app/model/unidade-responsavel.ts
@@ -1,0 +1,6 @@
+export class UnidadeResponsavel{
+  sigla: string = '';
+  nome: string = '';
+  constructor(){
+  }
+}