+++ client/src/app/services/index.ts
@@ -16,3 +16,4 @@ export * from './criterio-agregacao.service';
 export * from './version.service';
 export * from './usuario.service';
 export * from './consulta.service';
+export * from './polaridade.service';