+++ server/api/helpers/job.js
@@ -24,6 +24,8 @@ const config = {
   idleTimeoutMillis: 30000, // how long a client is allowed to remain idle before being closed
 };
 
+const schema = process.env.SCHEMA || config.schema;
+
 const pool = new pg.Pool(config);
 pool.on('error', function (err, client) {
   console.error('idle client error', err.message, err.stack);
@@ -36,11 +38,11 @@ exports.cron = ()=>{
       switch (item.codigo) {
         case 360:
           schedule.scheduleJob('42 * 1 * * *', ()=>{
-              var sql = `update dbesusgestor.tb_indicador set ds_ultima_atualizacao=maior_data
-              from (select a.co_seq_indicador, max(co_ano) as maior_data from dbesusgestor.tb_indicador a inner join ${config_param.schema_esusgestor}.${config_param.tabela_indicadores} b
+              var sql = `update ${schema}.tb_indicador set ds_ultima_atualizacao=maior_data
+              from (select a.co_seq_indicador, max(co_ano) as maior_data from ${schema}.tb_indicador a inner join ${config_param.schema}.${config_param.tabela_indicadores} b
               on a.co_seq_indicador=b.co_seq_indicador
               group by a.co_seq_indicador) as sq
-              where dbesusgestor.tb_indicador.co_seq_indicador = sq.co_seq_indicador
+              where ${schema}.tb_indicador.co_seq_indicador = sq.co_seq_indicador
               and co_periodicidade_atualizacao=${item.codigo};`;
               pool.query(sql,null, (err, result)=>{
                   if(err) {
@@ -53,11 +55,11 @@ exports.cron = ()=>{
             break;
           case 30:
             schedule.scheduleJob('42 * 2 * * *', ()=>{
-                var sql = `update dbesusgestor.tb_indicador set ds_ultima_atualizacao=maior_data
-                from (select a.co_seq_indicador, max(co_anomes) as maior_data from dbesusgestor.tb_indicador a inner join ${config_param.schema_esusgestor}.${config_param.tabela_indicadores} b
+                var sql = `update ${schema}.tb_indicador set ds_ultima_atualizacao=maior_data
+                from (select a.co_seq_indicador, max(co_anomes) as maior_data from ${schema}.tb_indicador a inner join ${config_param.schema}.${config_param.tabela_indicadores} b
                 on a.co_seq_indicador=b.co_seq_indicador
                 group by a.co_seq_indicador) as sq
-                where dbesusgestor.tb_indicador.co_seq_indicador = sq.co_seq_indicador
+                where ${schema}.tb_indicador.co_seq_indicador = sq.co_seq_indicador
                 and co_periodicidade_atualizacao=${item.codigo};`;
                 pool.query(sql,null, (err, result)=>{
                     if(err) {