+++ server/api/models/polaridade.js
@@ -3,25 +3,24 @@ const config_param = require('../helpers/config')();
 const schema = process.env.SCHEMA || config_param.schema;
 
 module.exports = function(sequelize, DataTypes) {
-  var UnidadeMedida = sequelize.define('UnidadeMedida', {
+  var Polaridade = sequelize.define('Polaridade', {
     codigo: {
         type: DataTypes.INTEGER,
         allowNull: false,
-        autoIncrement: true,
         primaryKey: true,
-        field: 'co_unidade_medida'
+        field: 'co_polaridade'
     },
     descricao: {
-        type: DataTypes.STRING,
-        field: 'ds_unidade_medida',
-        notEmpty: true,
+        type: DataTypes.STRING(50),
+        field: 'ds_polaridade',
         allowNull: false
-      }
+    }
   },{
     schema: schema,
     timestamps: false,
     freezeTableName: true,
-    tableName: 'tb_unidade_medida'
+    tableName: 'tb_polaridade'
   });
-  return UnidadeMedida;
+
+  return Polaridade;
 };