+++ server/api/models/unidade.js
@@ -1,6 +1,6 @@
 'use strict';
 const config_param = require('../helpers/config')();
-const schema = process.env.SCHEMA || config_param.schema_esusgestor;
+const schema = process.env.SCHEMA || config_param.schema;
 
 module.exports = function(sequelize, DataTypes) {
   var Unidade = sequelize.define('Unidade', {
@@ -65,7 +65,7 @@ module.exports = function(sequelize, DataTypes) {
     hierarchy: {
       foreignKey: 'unidade_pai',
       levelFieldName: 'nu_nivel',
-      throughSchema: 'dbesusgestor',
+      throughSchema: schema,
       throughTable: 'tb_unidade_hierarquia',
       throughForeignKey:'co_unidade_superior',
       throughKey: 'co_unidade'